Fixed the bit depth issue with making FITS cubes.
git-svn-id: https://aplpy.svn.sourceforge.net/svnroot/aplpy/trunk/src/APLpy@140 2f3f69f2-b874-4715-96d5-35985bcb53a6

+++ aplpy/montage.py
@@ -138,7 +138,7 @@ def make_rgb_cube(files,output):
     ny = int(header['NAXIS2'])
     
     # Generate emtpy cube
-    image_cube = np.zeros((len(files),ny,nx))
+    image_cube = np.zeros((len(files),ny,nx),dtype=np.float32)
     
     # Loop through files
     for i in range(len(files)):
